Event.reconfigures(), Event.alterations
Events can change the underlying World state now by reconfiguring Existents. The reconfigure() method of Event specifies what happens to change the world state. The alterations property (a list) of Event stores these changes. Every time an event is narrated, all of the chronologically earlier ones are checked and their "after" properties are applied to all appropriate Existents. Also, all of the chronologically later ones are checked and their "before" properties are applied. This assures that no matter what the order of narration, the World at that point seems to be in the right state.